added ebean.properties, removed create-line (hardcoded to "[tsigns]", rewrite ping and storing serverinfo, cleaned up a lot

+++ src/main/java/de/zh32/teleportsigns/PlayerListener.java
@@ -1,17 +1,11 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
 package de.zh32.teleportsigns;
 
 import de.zh32.teleportsigns.ping.Ping;
 import java.io.ByteArrayOutputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
-import java.util.Map.Entry;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
-import org.bukkit.Location;
 import org.bukkit.block.Sign;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -33,7 +27,7 @@ class PlayerListener implements Listener {
     
     @EventHandler
     private void onSignChange(SignChangeEvent e) {
-        if (e.getLine(0).equalsIgnoreCase(plugin.getConfig().getString("create-line")) && e.getPlayer().isOp()) {
+        if (e.getLine(0).equalsIgnoreCase("[tsigns]") && e.getPlayer().hasPermission("teleportsigns.create")) {
             TeleportSign ts = new TeleportSign();
             ts.setServer(e.getLine(1));
             ts.setLocation(e.getBlock().getLocation());
@@ -47,28 +41,26 @@ class PlayerListener implements Listener {
     private void onBlockBreak(BlockBreakEvent e) {
         if (e.getBlock().getState() instanceof Sign) {
             Sign s = (Sign) e.getBlock().getState();
-            if (s.getLine(0).equalsIgnoreCase(plugin.getConfig().getString("create-line")) || s.getLine(0).equalsIgnoreCase(ChatColor.translateAlternateColorCodes('&', plugin.getConfig().getString("first-line")))) {
-                TeleportSign ts = plugin.getDatabase().find(TeleportSign.class).where().
-                        eq("x", e.getBlock().getLocation().getX()).
-                        eq("y", e.getBlock().getLocation().getY()).
-                        eq("z", e.getBlock().getLocation().getZ()).findUnique();
-                if (ts != null) {
-                    plugin.getDatabase().delete(ts);
-                    plugin.loadSigns();
-                    e.getPlayer().sendMessage(ChatColor.GREEN + "Sign destroyed.");
-                }
+            TeleportSign ts = plugin.getDatabase().find(TeleportSign.class).where().
+                    eq("x", e.getBlock().getLocation().getX()).
+                    eq("y", e.getBlock().getLocation().getY()).
+                    eq("z", e.getBlock().getLocation().getZ()).findUnique();
+            if (ts != null && e.getPlayer().hasPermission("teleportsigns.destroy")) {
+                plugin.getDatabase().delete(ts);
+                plugin.loadSigns();
+                e.getPlayer().sendMessage(ChatColor.GREEN + "Sign destroyed.");
             }
         }
     }
     
     @EventHandler
     private void onClick(PlayerInteractEvent e) {
-        if (e.hasBlock() && e.getClickedBlock().getState() instanceof Sign && e.getAction() == Action.RIGHT_CLICK_BLOCK && e.getPlayer().hasPermission("lobby.teleport")) {
+        if (e.hasBlock() && e.getClickedBlock().getState() instanceof Sign && e.getAction() == Action.RIGHT_CLICK_BLOCK && e.getPlayer().hasPermission("teleportsigns.use")) {
             for (TeleportSign ts : plugin.signs) {
                 if (ts != null) {
                     if (ts.getLocation().equals(e.getClickedBlock().getLocation())) {
-                        if (Ping.getInstance().results.get(ts.getServer()) != null) {
-                            if (Ping.getInstance().results.get(ts.getServer()).isOnline()) {
+                        if (Ping.getInstance().serverinfos.get(ts.getServer()) != null) {
+                            if (Ping.getInstance().serverinfos.get(ts.getServer()).isOnline()) {
                                 ByteArrayOutputStream b = new ByteArrayOutputStream();
                                 DataOutputStream out = new DataOutputStream(b);
                                 try {
@@ -80,7 +72,7 @@ class PlayerListener implements Listener {
                                 e.getPlayer().sendPluginMessage(plugin, "BungeeCord", b.toByteArray());
                             }
                             else {
-                                e.getPlayer().sendMessage(ChatColor.translateAlternateColorCodes('&', plugin.getConfig().getString("offline-message")));
+                                e.getPlayer().sendMessage(ChatColor.translateAlternateColorCodes('&', plugin.config.offline_message));
                             }
                         }
                     }
